Add Dockerfile and .dockerignore
The Dockerfile specifies a container image, which only contains the
git-bug statically compiled binary and a preconfigured issue-repository.

- Statical compilation is required to run the git-bug binary on the
  scratch image, which reduces the size of the final container image.
- The included issue-repository allows the user of the image to try
  git-bug without the initial setup of a git repository and a git-bug
  idenentity.

The final container image will use the ca-certificates of the golang
image. This certificates are required to pull or push issues to/from
configured git-bug bridges.

Update installation and usage instructions for container image.
Also add a container image build status badge to README.