Add docker-compose setup for WebUI
This commit introduces a docker-compose file to reduce the setup for the
webui development.

The compose-file defines two containers. One for the GraphQL-backend and
another nodejs container for the development webui (dev-webui). With
this, it is no longer required to install go for compiling the git-bug
binary to start the dev-webui. A developer is only required to have
docker-compose installed to spin up the webui in development mode with
one command.

To support running the dev-webui inside a container, the proxy of the
dev-webui has to communicate with the backend-container instance.
Before this commit, the proxy listend to localhost, which prevented it
from communicating with the backend-container.
Changing the proxy URL to the backend-container IP or name, would be
fragile and certainly break the dev-webui for developers not using
docker-compose. So to support docker-compose gracefully, an
environmental variable is introduced which per default points to the
address of localhost and can be overwritten by docker-compose to the
backend-container instance.
As the proxy setting in package.json doesn't allow to retrieve it's
value from environmental variables, it has to be replaced by a more
capable solution - setupProxy.js.

NOTE: It would have been also possible to use the host-networkmode in
docker-compose. But platforms, which don't run containers natively,
might require additional steps to work.

The hosts node_modules directory is ignored and not mapped into the
webui-container. Instead the webui-container will download the
node_modules itself into a container-volume. This prevents the usage of
host dependent node_modules and should make the development environment
more reproducable.

The git-bug keyring directory is persistet via a container volume.
Furthermore by setting the KEYRING environmental variable in the .env-file, a
directory can be mounted to persist tokens at the container-host or provide
already persisted tokens from the host to the container.

Below are some additional files (changes):

- Add .env file to set the issue-repository for the GraphQL-backend
  service without changing the docker-compose.yml file.

- Reduce download size for docker build, by reducing npm install
  invokations through better cache use.