Add old-webui and use compose for bootstrap
The old-webui code is taken from the upstream branch.
The code is put here to run the old-webui in dev mode next to the
release version of the new webui. Running one webui in dev-mode is
required, as the release versions will lock the issue-repository.
The lock will then prevent the other webui from accessing the same
issue-repository.